requirements.txt
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
requests==2.32.3
requests-oauthlib==2.0.0
streamlit==1.49.1
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.4
matplotlib-inline==0.1.7
scipy==1.14.1
oauth2client==4.1.3
oauthlib==3.3.1
typing_extensions==4.14.0
typing-inspection==0.4.1
bcrypt
supabase
streamlit-oauth
pyjwt